% Generated by roxygen2 (4.0.2): do not edit by hand
\name{configureBBComp}
\alias{configureBBComp}
\title{Set global configuration options.}
\usage{
configureBBComp(history, logfilepath)
}
\arguments{
\item{history}{[\code{logical(1)}]\cr
Enable history? Enabling the history has the effect that each single function evaluation
is logged to local files. This means that this information is available even after
a program crash or power outage, possibly allowing an implementation to recover gracefully
and continue the optimization run. The downside is that the files may grow as large as 30 GB,
which may be an issue in some restricted environments. Also, the history information may not
be of value to all optimizers. This is why it can be turned off.}

\item{logfilepath}{[\code{character(1)}]\cr
This path is prepended to file names of log files.
This gives the user control over where to store these files.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
This call is optional, however, it can only occur before login.
}

