% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBudget}
\alias{getBudget}
\title{Return the evaluation budget for the current problem.}
\usage{
getBudget()
}
\value{
[\code{integer(1)}].
}
\description{
This is the total available budget, not the remaining budget (which can be obtained by
subtracting the result of the \code{\link{getEvaluations}} function).
Calling \code{\link{evalPoint}} more than this number of times on the current problem will fail.
}

