% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDimension}
\alias{getDimension}
\title{Return the dimension of the search space of the current problem.}
\usage{
getDimension()
}
\value{
[\code{integer(1)}].
}
\description{
Return the dimension of the search space of the current problem.
}

