% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getEvalHistory}
\alias{getEvalHistory}
\title{Evaluate the objective function.}
\usage{
getEvalHistory(indices)
}
\arguments{
\item{indices}{[\code{integer}]\cr
Time indices of the evaluations. If T queries to evaluate were already made then these
can be any number in the range 0 to T-1.
NOTE THAT THEREFORE AN INDEX IS ZERO-BASED UNLIKE USUALLY IN R!}
}
\value{
[\code{numeric(1)}]. Objective value at \code{point}.
}
\description{
Recover a previous evaluation from the log file.
If logging was disabled through \code{\link{configureBBComp}} or if the log file is removed
or corrupted then this function will fail.
}

