% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getEvaluations}
\alias{getEvaluations}
\title{Return the number of evaluations so far for the current problem.}
\usage{
getEvaluations()
}
\value{
[\code{integer(1)}].
}
\description{
Return the number of evaluations so far for the current problem.
}

