% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNumberOfProblems}
\alias{getNumberOfProblems}
\title{Return number of problems in selected track.}
\usage{
getNumberOfProblems()
}
\value{
[\code{integer(1)}].
}
\description{
Return the number of problems M in the current track. Valid problem IDs range from 0 to M-1.
}

