% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNumberOfTracks}
\alias{getNumberOfTracks}
\title{Return number of tracks.}
\usage{
getNumberOfTracks()
}
\value{
[\code{integer(1)}].
}
\description{
Return the number N of tracks available to the user.
Valid track indices range from 0 to N-1.
}

