% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTrackName}
\alias{getTrackName}
\title{Return the name of the track.}
\usage{
getTrackName(index)
}
\arguments{
\item{index}{[\code{integer(1)}]\cr
Index of selected track.}
}
\value{
[\code{character(1)}].
}
\description{
Tracks are identified by their name, not by their index (which may change depending on user rights).
}

