% Generated by roxygen2 (4.0.2): do not edit by hand
\name{login}
\alias{login}
\title{Login on server.}
\usage{
login(username, password)
}
\arguments{
\item{username}{[\code{character(1)}]\cr
Username for login.}

\item{password}{[\code{character(1)}]\cr
Password for login.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
This call results in communication with the competition server.
It validates the user's credentials. Multiple concurrent logins with the same credentials are
possible to allow for parallel evaluation of problems on different cores/machines,
but not from multiple threads within the same process.
The account becomes a persistent property of the library state; subsequent function calls refer
to this account.
}

