% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setProblem}
\alias{setProblem}
\title{Select a problem within the chosen track for optimization.

Subsequent function calls refer to this problem.}
\usage{
setProblem(id)
}
\arguments{
\item{id}{[\code{integer(1)}]\cr
ID of selected problem.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
Select a problem within the chosen track for optimization.

Subsequent function calls refer to this problem.
}

