% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setTrack}
\alias{setTrack}
\title{Set a track.}
\usage{
setTrack(track)
}
\arguments{
\item{track}{[\code{character(1)}]\cr
Name of selected track.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
Select a track. Subsequent function calls refer to this track.
}

